+++ src/lib/task-generators/clean.ts
@@ -1,4 +1,4 @@
-import * as del from "del";
+import del from "del";
 import { Gulp } from "gulp";
 import { Minimatch } from "minimatch";
 import {posix as path } from "path";